+++ packages/template-retail-react-app/app/pages/checkout/index.jsx
@@ -119,6 +119,9 @@ const Checkout = () => {
 
     return (
         <Box background="gray.50" flex="1">
+            <Heading as="h1" fontSize="2xl" mb={6} textAlign="center">
+                <FormattedMessage defaultMessage="Checkout" id="checkout.title.checkout" />
+            </Heading>
             <Container
                 data-testid="sf-checkout-container"
                 maxWidth="container.xl"